case in user22 commit

+++ app/Http/Controllers/api/user/UsCaseController.php
@@ -22,7 +22,7 @@ class UsCaseController extends Controller
             'user_id',
             'donationtype_id',
             'category_id'
-            )->with('category','donationtype','user')->get();
+            )->with('category','donationtype','user')->where('status','published')->get();
 
         $response = [
             'message'=>'All cases',
@@ -45,13 +45,10 @@ class UsCaseController extends Controller
             'user_id',
             'donationtype_id',
             'category_id'
-            )->where('id',$id)->first();
+            )->with('category','donationtype','user')->where('id',$id)->first();
         $response = [
             'message'=>'specific case with id',
-            'case' => $casee,
-            'category'=>$casee->category,
-            'donationtype'=>$casee->donationtype,
-            'user'=>$casee->user
+            'case' => $casee
         ];
         return response($response,201);
     }
@@ -63,10 +60,9 @@ class UsCaseController extends Controller
             'description_en' => 'string|max:500',
             'description_ar' => 'string|max:500',
             'image' => 'image|max:2048',
-            'Donantiontype_id' =>'required|exists:donationtypes,id',
-            'Category_id' =>'required|exists:categories,id',
+            'donationtype_id' =>'required|exists:donationtypes,id',
+            'category_id' =>'required|exists:categories,id',
             'initial_amount'=>'required|numeric',
-            'status'=>'required|in:pending,accepted,published,rejected'
         ]);
 
         if($request->file('image')){
@@ -82,11 +78,13 @@ class UsCaseController extends Controller
             'description_en'=> $request->description_en,
             'description_ar'=> $request->description_ar,
             'image' => $image,
-            'Donantiontype_id'=> $request->Donantiontype_id,
-            'Category_id'=> $request->Category_id,
+            'donationtype_id'=> $request->donationtype_id,
+            'category_id'=> $request->category_id,
             'initial_amount'=>$request->initial_amount,
-            'status'=>$request->status,
-            'User_id'=>1
+            'paied_amount'=>0,
+            'remaining_amount'=>0,
+            'user_id'=>$request->user->id,
+            'status'=>'pending'
         ]);
 
         $response = [
@@ -99,39 +97,55 @@ class UsCaseController extends Controller
     public function update(Request $request,$id){
         $casee=Casee::find($id);
 
+        if($casee->status=='pending'){
+            $request->validate([
+                'name_en' => 'required|string|max:200',
+                'name_ar' => 'required|string|max:200',
+                'description_en' => 'string|max:500',
+                'description_ar' => 'string|max:500',
+                'image' => 'image|max:2048',
+                'donationtype_id' =>'required|exists:donationtypes,id',
+                'category_id' =>'required|exists:categories,id',
+                'initial_amount'=>'required|numeric',
+                'status'=>'required|in:pending'
+            ]);
+        }else{
+
         $request->validate([
             'name_en' => 'required|string|max:200',
             'name_ar' => 'required|string|max:200',
             'description_en' => 'string|max:500',
             'description_ar' => 'string|max:500',
             'image' => 'image|max:2048',
-            'Donantiontype_id' =>'required|exists:donationtypes,id',
-            'Category_id' =>'required|exists:categories,id',
+            'donationtype_id' =>'required|exists:donationtypes,id',
+            'category_id' =>'required|exists:categories,id',
             'initial_amount'=>'required|numeric',
-            'status'=>'required|in:pending,accepted,published,rejected'
-        ]);
+            'status'=>'required|in:accepted,published'
+        ]);}
 
         if($request->file('image')){
             $image_path = $request->file('image')->store('api/casees','public');
-            $casee->image = asset('storage/'.$image_path);
-            $casee->save();
+            $image=asset('storage/'.$image_path);
+        }else{
+            $image=$casee->image;
         }
-        
+
+
+
         $casee->update([
             'name_en' => $request->name_en,
             'name_ar'=> $request->name_ar,
             'description_en'=> $request->description_en,
             'description_ar'=> $request->description_ar,
-            'image' => asset('storage/'.$image_path),
-            'Donantiontype_id'=> $request->Donantiontype_id,
-            'Category_id'=> $request->Category_id,
+            'image' => $image,
+            'donationtype_id'=> $request->donationtype_id,
+            'category_id'=> $request->category_id,
             'initial_amount'=>$request->initial_amount,
-            'status'=>$request->status,
-            'User_id'=>1
+            'user_id'=>$request->user->id,
+            'status'=>$request->status
                 ]);
         
         
-
         $response = [
             'message'=>'case updated successfully',
             'case' => $casee
@@ -139,12 +153,20 @@ class UsCaseController extends Controller
         return response($response,201);
     }
 
-    public function destroy($id){
+    public function destroy(Request $request,$id){
+
         $casee = Casee::findOrFail($id);
-        $casee->delete();
+        if($casee->user_id==$request->user->id){
+            $casee->delete();
         $response = [
             'message'=>'case deleted successfully',
         ];
-        return response($response,201);
+        
+        }else{
+            $response = [
+                'message'=>'can not be deleted ',
+            ];
+        }
+            return response($response,201);
     }
 }
